requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
aiosqlite==0.20.0
httpx==0.28.1
langfuse==2.57.3
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.0
sse-starlette==2.2.1
python-multipart==0.0.20
